pyproject.toml
[project]
name = "pr-1-rag"
version = "0.1.0"
description = "rag agent"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11,<3.15"
dependencies = [
    "sentence-transformers (>=5.4.1,<6.0.0)",
    "pymupdf (>=1.27.2.2,<2.0.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "groq (>=1.1.2,<2.0.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
